Update sketch.js
Aggressive pivot because I can't figure out my original plan and this still has the essence of my original plan. Instead of adding things to an array, you can just select the intensity bubble and write a task you want to complete. (unless teach gets back to me soon, this is what we're going with)